AiAgentGame/pyproject.toml
[project]
name = "ai-agent-game"
version = "0.1.0"
description = "AI Agent Game Creator - Multi-agent system for automatic game development"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",
    "langsmith>=0.1.0",
    "watchdog>=3.0.0",
    "pyyaml>=6.0.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "pydub>=0.25.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "pygame>=2.5.0",
    "pyxel>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
